Rewrite HTTP->HTTPS redirection
See the comment in sites-available/10-plaintext-catchall.conf for the
details, but basically that file now magically handles all traffic on
port 80. It automatically forces HTTPS, except in the case of an onion
request, in which case it forwards it on to the normal clearnet handler.

Still TODO at this stage:

* Expand onion services to cover all subdomains
* Log if something comes from an onion service maybe? I dunno
* Fix null.strugee.net and associated logic, since that broke
* Try and figure out a way to get rid of the need to proxy requests to
  ourselves, though I don't hold out much hope for this one :(